Move tests from cypress -> puppeteer
