+++ puppeteer/atiAnalytics/assertions/scrollablePromo.js
